Assigns a sign change between directions to a transcript/node

Assigns a sign change between directions to a transcript/node

Usage

computeSign(y, x, minP = 0.7, pc = 5)

Arguments

y

SummarizedExperiment containing the scaled inferential replicates

x

character which is the column in colData that contains the condition information

minP

float value b/w 0 and 1, denoting how much proportion of the inferential replicates should hold the same sign change

pc

numeric, pseudocount, default is 5

Value

numeric vector, of length equal to the number of rows in y, with 1 indicating positive sign change, -1 negative sign change and 0 unsure