mehenDi: Differential testing on the forest of trees output form TreeTerminus

This package provides a tree based differential testing for RNA-Seq data, where the tree is obtained from TreeTerminus. The candidate nodes can consist of both leaves and inner nodes, with inner nodes comprising a susbet of transcripts that exhibit high uncertainty.

See also

Author

Maintainer: Noor Pratap Singh noorpratap@gmail.com

Authors:

Michael Love michaelisaiahlove@gmail.com [contributor]

Rob Pato rob@cs.umd.edu [contributor]