Function is deprecated in favor of mehenDi

Function is deprecated in favor of mehenDi

Usage

trenDi(
  tse,
  x,
  pvalues,
  minP = 0.7,
  mIRVThresh = 0.4,
  alpha = 0.01,
  pThresh = NULL,
  cores = 1
)

Arguments

tse

TreeSummarizedExperiment which contains the scaled inferential replicates

x

character, the name of the condition variable. A factor for two group analysis

pvalues

numeric, pvalues for all the nodes in the tree

minP

numeric, value betweem 0 and 1, the proportion of the total inferential replicates which should have the same sign change between conditions (default 0.7)

mIRVThresh

numeric, minimum meanInfRV that a node should have for it to be considered for aggregation (default 0.4)

alpha

numeric, the rate for the BH correction on the leaves which will be the pvalue threshold for deeming a node significant, wont be used if pThresh is set to some value

pThresh

numeric, pvalue threshold for climbing, default is NULL since pThresh is computed automatically based on alpha, if not set to NULL then this pvalue threshold will used for deeming a node significant

cores

numeric, the number of cores that will be used during parallelization